planificator v2/requirements.txt
Flask==2.3.3
numpy==1.24.3
pandas==2.0.3
openpyxl==3.1.2
python-dateutil==2.8.2
python-docx
rapidfuzz
